requirements.txt
pandas==2.2.2
google-api-python-client==2.129.0
mysql-connector-python==8.4.0
streamlit==1.34.0
toml==0.10.2
